pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tech-pulse"
version = "0.1.0"
description = "AI-powered tech news intelligence pipeline"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "anthropic>=0.40.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.7.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
